+++ openselfsup/datasets/data_sources/__init__.py
@@ -2,4 +2,6 @@ from .cifar import Cifar10, Cifar100
 from .image_list import ImageList
 from .image_list_multihead import ImageListMultihead
 from .imagenet import ImageNet
+from .sen12ms import Sen12MS
+from .sen12ms_image_list import Sen12MSImageList
 from .places205 import Places205
