+++ openselfsup/datasets/loader/build_loader.py
@@ -90,7 +90,6 @@ def worker_init_fn(seed):
     np.random.seed(seed)
     random.seed(seed)
 
-
 class PrefetchLoader:
     """
     A data loader wrapper for prefetching data
